Shiva/DailyJournal-QuantumLeap-865903-2099626-919237/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
mysql-connector-python
python-jose
pydantic
python-dotenv
